Added a FileUtils and extracted the methods from the test into it
